r1.1 update and added README
Version r1.0 of my code works perfectly fine with chests, shulkers and barrels, but the other containers still had the Full Inventory ingredient deletion glitch.
Now it detects the type of container the item frame is place on and changes the maximum number of empty slots accordingly.
Added README explaining how to use the datapack and every change I made to McRadane's code.